src/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
psycopg2==2.7.6.1
whitenoise==4.1
uwsgi
